+++ src/client/repl/replUtils.ts
@@ -1,8 +1,21 @@
-import { TextEditor } from 'vscode';
+import {
+    NotebookCellData,
+    NotebookCellKind,
+    NotebookDocument,
+    NotebookEdit,
+    TextEditor,
+    workspace,
+    WorkspaceEdit,
+} from 'vscode';
 import { getActiveResource } from '../common/vscodeApis/windowApis';
 import { getConfiguration } from '../common/vscodeApis/workspaceApis';
 import { getMultiLineSelectionText, getSingleLineSelectionText } from '../terminals/codeExecution/helper';
 
+/**
+ * Function that returns selected text to execute in the REPL.
+ * @param textEditor
+ * @returns code - Code to execute in the REPL.
+ */
 export async function getSelectedTextToExecute(textEditor: TextEditor): Promise<string | undefined> {
     if (!textEditor) {
         return undefined;
@@ -22,8 +35,28 @@ export async function getSelectedTextToExecute(textEditor: TextEditor): Promise<
     return code;
 }
 
+/**
+ * Function that returns user's Native REPL setting.
+ * @returns boolean - True if sendToNativeREPL setting is enabled, False otherwise.
+ */
 export function getSendToNativeREPLSetting(): boolean {
     const uri = getActiveResource();
     const configuration = getConfiguration('python', uri);
     return configuration.get<boolean>('REPL.sendToNativeREPL', false);
+}
+
+/**
+ * Function that adds cell to notebook.
+ * This function will only get called when notebook document is defined.
+ * @param code
+ *
+ */
+export async function addCellToNotebook(notebookDocument: NotebookDocument, code: string): Promise<void> {
+    const notebookCellData = new NotebookCellData(NotebookCellKind.Code, code as string, 'python');
+    const { cellCount } = notebookDocument!;
+    // Add new cell to interactive window document
+    const notebookEdit = NotebookEdit.insertCells(cellCount, [notebookCellData]);
+    const workspaceEdit = new WorkspaceEdit();
+    workspaceEdit.set(notebookDocument!.uri, [notebookEdit]);
+    await workspace.applyEdit(workspaceEdit);
 }
